+++ .gitignore
@@ -169,4 +169,4 @@ recipes/**/outputs/
 
 # Client recipe connection configs — these hold a Snowflake PAT. The checked-in
 # config.json.template is not matched (it does not end in .json).
-arctic_platform/client/recipes/*.json
+recipes/*.json
