+++ recipes/sft/standalone/requirements.txt
@@ -1,0 +1,17 @@
+# Everything this recipe needs on top of the client itself.
+#
+#   uv pip install -e '.[cortex]'
+#   uv pip install -r recipes/sft/standalone/requirements.txt
+
+# Config entrypoints (`config=... lora_rank=32` on the command line).
+chz
+
+# Chat dataset loading; also pulled in by tinker-cookbook.
+datasets
+
+# Renderers (chat templates + supervised weights) and ml_log metric logging.
+# SFT needs no math grading, so the [math-rl] extra is not required here.
+tinker-cookbook @ git+https://github.com/thinking-machines-lab/tinker-cookbook.git@nightly
+
+# Optional: mirror metrics to Weights & Biases (set wandb_project=... to enable).
+wandb
